requirements-dev.txt
# Development dependencies — install with: pip install -r requirements-dev.txt
# Runtime deps live in requirements.txt (kept lean for the production image).

# Tests
pytest==8.3.4
pytest-asyncio==0.25.0

# Lint + format
ruff==0.8.4

# Static type check
mypy==1.14.0
boto3-stubs==1.35.86
types-requests==2.32.0.20241016

# Pre-commit hooks
pre-commit==4.0.1
